+++ mm/oom/oom.c
@@ -18,27 +18,33 @@
 
 volatile sig_atomic_t keep_going = 1;
 
-size_t mem_size = 0;
-enum {
+enum ops_type {
 	OP_GLIBC = 1,
 	OP_MMAP_ANON,
 	OP_MMAP_FILE,
-} op_type = OP_GLIBC; /* default use glibc */
-int oom_adj;
-int oom_score_adj;
-bool flag_popen =
-#if !defined(POPEN)
-	false;
-#pragma message("without popen(3)")
-#else
-	true;
-#pragma message("with popen(3)")
-#endif
-int verbose = false;
-unsigned long rate_limit = 9999999999;
+};
+
+struct {
+	enum ops_type ops_type;
+	size_t mem_size;
+	int oom_adj;
+	int oom_score_adj;
+	bool flag_popen;
+	unsigned long rate_limit;
+	size_t limit_in_bytes;
+	int verbose;
+} env = {
+	.ops_type = OP_GLIBC,	/* default use glibc */
+	.mem_size = 0,
+	.oom_adj = 0,
+	.oom_score_adj = 0,
+	.flag_popen = false,
+	.rate_limit = 9999999999,
 #ifdef HAVE_LIBCGROUP
-size_t limit_in_bytes = 0;
+	.limit_in_bytes = 0,
 #endif
+	.verbose = false,
+};
 
 const char argp_prog_doc[] =
 	"USAGE: [-p] [-s <size>] [-a <oom_adj>] [-c <oom_score_adj>] [-v|--verbose]\n"
@@ -67,38 +73,38 @@ static error_t parse_arg(int key, char *arg, struct argp_state *state)
 	switch (key) {
 	case 'e':
 		if (!strcmp(arg, "glibc"))
-			op_type = OP_GLIBC;
+			env.ops_type = OP_GLIBC;
 		else if (!strcmp(arg, "mmap-anon"))
-			op_type = OP_MMAP_ANON;
+			env.ops_type = OP_MMAP_ANON;
 		else if (!strcmp(arg, "mmap-file"))
-			op_type = OP_MMAP_FILE;
+			env.ops_type = OP_MMAP_FILE;
 		else {
 			fprintf(stderr, "ERROR: operation only glibc, mmap-anon, mmap-file");
 			exit(EXIT_FAILURE);
 		}
 		break;
 	case 's':
-		mem_size = str2size(arg);
+		env.mem_size = str2size(arg);
 		break;
 #ifdef HAVE_LIBCGROUP
 	case 'M':
-		limit_in_bytes = str2size(arg);
+		env.limit_in_bytes = str2size(arg);
 		break;
 #endif
 	case 'r':
-		rate_limit = str2size(arg);
+		env.rate_limit = str2size(arg);
 		break;
 	case 'a':
-		oom_adj = atoi(arg);
+		env.oom_adj = atoi(arg);
 		break;
 	case 'c':
-		oom_score_adj = atoi(arg);
+		env.oom_score_adj = atoi(arg);
 		break;
 	case 'p':
-		flag_popen = true;
+		env.flag_popen = true;
 		break;
 	case 'v':
-		verbose = true;
+		env.verbose = true;
 		break;
 	case ARGP_KEY_ARG:
 		argp_usage(state);
@@ -153,7 +159,7 @@ struct oom_operations {
 	const char *name;
 	size_t total_size;
 	void *(*alloc)(size_t size);
-	void (*pagefault)(void *mem, size_t size, bool verbose);
+	void (*pagefault)(void *mem, size_t size, bool pf_verbose);
 	void (*free)(void *mem, size_t size);
 };
 
@@ -178,7 +184,7 @@ void mmap_anon_free(void *mem, size_t size)
 	munmap(mem, size);
 }
 
-void default_pagefault(void *mem, size_t size, bool verbose)
+void default_pagefault(void *mem, size_t size, bool pf_verbose)
 {
 	size_t i;
 	size_t pf_size = 0;
@@ -188,7 +194,7 @@ void default_pagefault(void *mem, size_t size, bool verbose)
 		((char *)mem)[i] = 'a';
 
 		pf_size += pagesize;
-		if (verbose) {
+		if (pf_verbose) {
 			BACK_PRINTF("Pagefault %ld B (%ld KiB, %ld MiB),"
 					" oom_score %d, adj %d, score_adj %d",
 					pf_size, pf_size / KB, pf_size / MB,
@@ -258,7 +264,7 @@ void hold_mem(struct oom_operations *ops)
 
 	while (keep_going) {
 		ops->pagefault(mem, size, true);
-		if (flag_popen)
+		if (env.flag_popen)
 			test_popen();
 		sleep(1);
 	}
@@ -299,10 +305,10 @@ void try_oom(struct oom_operations *ops)
 
 		rate_Mps = cal_rate_size * 1.0f / (usecs() - start);
 
-		if (flag_popen)
+		if (env.flag_popen)
 			test_popen();
 
-		if (verbose) {
+		if (env.verbose) {
 			BACK_PRINTF("allocated %ld B (%ld MiB, %ld GiB), %.2lf MiB/s,"
 				    " oom_score %d, adj %d, score_adj %d",
 				    ops->total_size, ops->total_size / 1024 / 1024,
@@ -313,7 +319,7 @@ void try_oom(struct oom_operations *ops)
 				    get_oom_score_adj(getpid()));
 		}
 		/* Limit the allocate rate */
-		if (rate_Mps > (rate_limit / 1024.0f / 1024.0f))
+		if (rate_Mps > (env.rate_limit / 1024.0f / 1024.0f))
 			usleep(5000);
 	}
 
@@ -327,8 +333,9 @@ static struct cgroup_controller *memcg = NULL;
 int memcg_limit(void)
 {
 	int err = 0;
+	size_t limit_in_bytes = 0;
 
-	if (limit_in_bytes == 0)
+	if (env.limit_in_bytes == 0)
 		return 0;
 
 	err = cgroup_init();
@@ -359,7 +366,7 @@ int memcg_limit(void)
 # define MEMORY_SWAP_LIMIT	"memory.swap.max"
 #endif
 
-	err = cgroup_set_value_uint64(memcg, MEMORY_LIMIT, limit_in_bytes);
+	err = cgroup_set_value_uint64(memcg, MEMORY_LIMIT, env.limit_in_bytes);
 	if (err) {
 		fprintf(stderr, "Error setting memory limit: %s, %s\n", CGRP_NAME,
 			cgroup_strerror(cgroup_get_last_errno()));
@@ -374,9 +381,8 @@ int memcg_limit(void)
 		goto free;
 	}
 
-	limit_in_bytes = 0;
 	cgroup_get_value_uint64(memcg, MEMORY_LIMIT, &limit_in_bytes);
-	if (verbose) {
+	if (env.verbose) {
 		printf("Set "MEMORY_LIMIT" to %ld MB\n", limit_in_bytes / 1024 / 1024);
 	}
 
@@ -387,7 +393,7 @@ int memcg_limit(void)
 		goto free;
 	}
 
-	if (verbose) {
+	if (env.verbose) {
 		printf("Attaching process to cgroup.\n");
 	}
 	err = cgroup_attach_task_pid(cgrp, getpid());
@@ -436,21 +442,21 @@ int main(int argc, char *argv[])
 
 	mlockall(MCL_CURRENT);
 
-	if (oom_adj) {
-		printf("set oom_adj to %d\n", oom_adj);
-		err = set_oom_adj(getpid(), oom_adj);
+	if (env.oom_adj) {
+		printf("set oom_adj to %d\n", env.oom_adj);
+		err = set_oom_adj(getpid(), env.oom_adj);
 		if (err != 0)
 			return err;
 	}
 
-	if (oom_score_adj) {
-		printf("set oom_score_adj to %d\n", oom_score_adj);
-		err = set_oom_score_adj(getpid(), oom_score_adj);
+	if (env.oom_score_adj) {
+		printf("set oom_score_adj to %d\n", env.oom_score_adj);
+		err = set_oom_score_adj(getpid(), env.oom_score_adj);
 		if (err != 0)
 			return err;
 	}
 
-	if (verbose) {
+	if (env.verbose) {
 		char comm[64];
 		printf("pid %d, comm %s\n", getpid(), proc_comm(comm, sizeof(comm)));
 		printf("oom_adj %d\n", get_oom_adj(getpid()));
@@ -459,7 +465,7 @@ int main(int argc, char *argv[])
 	}
 
 	struct oom_operations *ops;
-	switch (op_type) {
+	switch (env.ops_type) {
 	case OP_GLIBC:
 		ops = &glibc_ops;
 		break;
@@ -472,12 +478,13 @@ int main(int argc, char *argv[])
 		break;
 	}
 
-	if (verbose) {
+	if (env.verbose) {
 		printf("Use OOM operator %s\n", ops->name);
+		printf("rate_limit %ld\n", env.rate_limit);
 	}
 
-	if (mem_size) {
-		ops->total_size = mem_size;
+	if (env.mem_size) {
+		ops->total_size = env.mem_size;
 		hold_mem(ops);
 	} else {
 		try_oom(ops);